Extract the logic into an extension
